function naming and redundancy
